HSTTWO-1631: introduce Channel storage
A Channel is associated with a mount.  Each mount can have a hst:channelid
property that identifies the channel.  Each channel should also have a
blueprint ID.  Each mount that belong to the same channel must have the same
blueprint ID.

The ChannelManager is made available as an extension to the HstManager
interface.  It provides methods to retrieve, create and modify channels.
Channels themselves are Serializable, allowing clients to store a copy while
they are editing it.

The first hint of a blueprint implementation is also here, though this is still
work-in-progress.