HSTTWO-1631: use same annotation mechanism for channel info and (component) parameter info
A blueprint can specify a channel info class, an interface with annotations on the site (hst core) classpath.  An instance of this class is made available on a mount, a channel itself exposes it as a set of HstPropertyDefinitions.

The existing Parameter class has been split (backwards compatible) into a number of different annotations.  The idea is that the user interface can be extended by additional annotations, without the need for extending the Parameter class any further.  The two existing extensions (Color and DocumentLink) have been added to the API.  The ParameterType that was in the template composer has been moved into the API.  The recommendation is to not use it though.  Instead, rely on the method return type and annotate when necessary.

The existing template composer code has been modified to deal with both legacy and current methods of annotating parameters.