HSTTWO-1811: tighter integration between channel manager service and hst manager
The channel manager now uses the virtual hosts from the hst manager to retrieve
mount information.  It does still need to create mounts itself (there is no
infrastructure yet to move this functionality into the hst manager).

The hst manager initializes the channel manager with it's virtual hosts when
they are rebuilt.  Since this needs to be synchronized, the hst manager has
been put in control of the synchronization, calling into the channel manager
when it is safe to modify the mount services.

The MutableChannelManager and MutableMount interfaces have been introduced to
keep the managers decoupled, making the information that needs to be present
for each component, explicit.