CMS7-5669: removed the need for a superfluous site node in a blueprint, removed 'mandatory' requirement for hst:configuration node in a blueprint
A blueprint no longer must have an hst:site node or hst:configuration node (although it still needs at least one of the two, otherwise its unclear which configuration to use for a new channel). The property 'hst:configurationpath' is also no longer set, as the HST derives it from convention. The channel manager now throws an exception if neither a content root path is set in the channel nor a content prototype is included in a blueprint.