CMS7-5706: close synchronization gap
When an event would received that cannot be processed, the complete virtual host configuration would be rebuilt.  The retrieval of the virtual hosts did not store a reference on the stack before leaving the monitor.  SInce releasing a contended lock typically schedules another (blocked) thread, the returned virtual hosts value could be null.