HSTTWO-2022: added REST service that return all active HTTP sessions.
The added REST service adds a mount at 127.0.0.1/_bt. The logic to add such a custom mount has been extracted from the hst-cms-rest module and generalized in the CustomMountAndVirtualHostAugmenter.

The HTTP sessions are tracked by the BehavioralDataHttpSessionStore whenever behavioral data is read, and stored in an HttpSessionRegistry. The implementation of this registry stores the sessions in weak references and automatically cleans up sessions that are no longer referenced.