HSTTWO-2199 use the canonical path as entry path for generating pages. This part was done by changing in MountService the
contentPath = hstSiteNodeForMount.getContentPath();

TO 
contentPath = hstSiteNodeForMount.getCanonicalContentPath();

Furthermore, this commit contains the removal of the HstContextWhereClauseComputer, as we can now just use the identifiers of the nodes (scopes) to search below. As a result, the virtualizers also have become redundant, which resulted in removing the HstContextualizeException and HstCtxWhereFilter as well. Nice....clean up