HSTTWO-2497: support @FieldGroupList and @FieldGroup annotations on component parameter info interfaces
Component parameters can now be grouped by including them in the same @FieldGroup annotation. The order of the parameters in the @FieldGroup annotation determines the order in which they are rendered in the template composer. Each group can optionally have a i18n'd title that will be rendered above the group when present.

This change is backwards compatible with older versions of the CMS, since these will ignore the additional field 'groupLabel' in the REST response for the parameters of a container item.