HSTTWO-2619: put the correct CMS user id on the HTTP session
With overlayed JCR sessions, the user id becomes a concatenation of the user ids of the overlayed sessions.
This combined 'user name' was put on the HTTP session. However, various code relies on the fact that this 
cms user name session attribute corresponds to actual CMS user names. For example, the CmsComponentWindowResponseAppender 
checks at line 94 whether the current user matches the user that locked a container in the template composer. 
This check was no longer valid, causing a container to become locked for the current user when components 
in it where reordered.