HSTTWO-2850 Convert validation exceptions to client exceptions
Instead of catching illegal argument exceptions and others, a
new ClientException has been introduced. Whenever a validation
error occurs that requires error feedback in the front-end a
ClientException should be thrown containing a ClientError (this
is the error code), an error message and a list of message parameters.

TODO: Properly define the errors that require feedback in the front-end.