HSTTWO-3149 implemented a cleaner separation and added better javadoc specification what defines a container resource and a non container resource: A container resource link is a resource relative to the webapp (context path) while a non container resource link is relative to the (sub)mount. Note that container resource links for root mount are the same a non container resource links. The difference is only for submounts as they include for non container resource links also the sub mountpath (eg /fr). Sitemap items that are marked to serve container resources (hst:containerresource = true) are only allowed below /hst:default/hst:sitemap. The items *are* inherited and can be overridden in submounts, but in general is unlikely that it will be done. Examples of what container resources are: 1) static webapp  files like css, js, jsp, gif, jpeg etc 2) web resources (the new ones loaded from the repository) 3) non publishable assets and gallery items
There are two different ways a URL can be a container resource: The first one is non-hst-model related but spring managed through  HstManager#isHstFilterExcludedPath. When a URL is marked by prefix or suffix to match exclusion, the HST will invoke the next filter in the servlet filter chain. Also linkrewriting will when a pathInfo matches one of the exclusion paths, mark the HstLink as a container resource. The other way that a URL / pathInfo can be a container resource, is when it matches a hst:sitemapitem that is marked with hst:containerresource = true. This second category of container resources do when they hit the webapp get their request matched against the hst hosts and sitemap, and then get processed by the 'PlainFilterChainInvokingPipeline' pipeline