HSTTWO-3184 Expose lock method via Spring bean
For targeting we need to be able to lock component items from within a REST endpoint. Currently this method is only exposed via the HST Rest endpoint method "lock" in the ContainerItemComponentResource. By factoring out the relevant method into a service we can expose it as a Spring bean and inject it in targeting.