HSTTWO-3201 make comment attribute generation plug-able for CMS response appender
I introduced the AttributeContributor interface and moved the attribute generation
part from the response appender into the CmsComponentAttributeContributer (which
implements this interface). The resource appender now iterates over a list of 
attribute contributors to produce the attribute map needed to generate the HTML
comment. By exposing the list as a Spring Bean it is possible to add custom
attribute contributors.