HSTTWO-3434 support cascading hst inheritance : From now on, an hst configuration setup as follows: + hst:hst    + hst:configurations       + common       + corporate           - hst:inheritsfrom = ../common           + hst:sitemenus           + hst:workspace               + hst:sitemenus       + subsite           - hst:inheritsfrom = ../corporate
makes sure that 'subsite' inherits from 'corporate' *and* from 'common'. Note that for the above setup, 'subsite' does not yet inherit the hst:workspace from 'corporate'. This can be achieve by:
      + subsite 
          - hst:inheritsfrom = ../corporate, ../corporate/hst:workspace
          
Note that if the configuration would had been:
+ hst:hst
   + hst:configurations
      + common
      + corporate
          - hst:inheritsfrom = ../common, ../common/hst:workspace
          + hst:sitemenus
          + hst:workspace
              + hst:sitemenus
      + subsite 
          - hst:inheritsfrom = ../corporate

in this case, subsite does not inherit the hst:workspace from 'corporate' but it does inherit the 'hst:workspace' from 'common' because 'corporate' inherits from ../common, ../common/hst:workspace.

Another detail: In case the setup is:

+ hst:hst
   + hst:configurations
      + common
      + corporate
          - hst:inheritsfrom = ../common, ../common/hst:workspace
          + hst:sitemenus
          + hst:workspace
              + hst:sitemenus
      + subsite 
          - hst:inheritsfrom = ../corporate, ../corporate/hst:workspace
          
then 'subsite' inherits in the following order the configuration:
1) corporate
2) common
3) common/hst:workspace  
4) corporate/hst:workspace
this is because the *first* inheritacance (../corporate) is first completely inherited. Also note that the above configuration is not very likely to really occur in practice. 
Note that if the configuration would had been:
+ hst:hst
   + hst:configurations
      + common
      + corporate
          - hst:inheritsfrom = ../common, ../common/hst:workspace
          + hst:sitemenus
          + hst:workspace
              + hst:sitemenus
      + subsite 
          - hst:inheritsfrom = ../corporate/hst:workspace, ../corporate

then the order of inheritance would had been:
1) corporate/hst:workspace   
2) corporate
3) common
4) common/hst:workspace
The order of inheritance is important when it comes to 'shielding' configuration nodes in case of same node names 
        
All the above scenarios have been added in unit tests