HSTTWO-3648 [Backport 3.1.3] Make sure that URL are created correctly for Mount that have ismapped=false
The bug was a regression which was a result of the work on the container resources.
Version 10 has the same problem, hence I will backport to 10 as well

Added a unit test that did break before the fix and passes after the fix. Should also avoid future mistakes

(cherry picked from commit 5fbc7bfa54f0e035794228973e6f210840c13599)