HSTTWO-3675 when a response state gets flushed, also flush possible set headers of unflushed child response states
I chose to implement this in the HstServletResponseState to be able to have the flushing of child states in
a single place. My first implementation was a helper method in HstResponseImpl but this meant that for the
root component window I had to explicitly call the  helper method in HstResponseImpl as well.

With the final implementation, all logic resides in the HstServletResponseState. I had to give the HstServletResponseState an extra constructor argument because it requires a reference to the backing
component window.