HSTTWO-3675 simplify HstComponentWindow / HstResponseState coupling
By adding a HstComponentWindow#setResponseState we can in the HstServletResponseState
bind the window directly to the response state instead of having to invoke a separate method
on HstComponentWindowImpl (and casting).

Also, I could get easily rid of the NoopResponseState in the Aggregation valve : It is even better to *not* use
a NoopResponseState : Namely, in case of 'component rendering that is marked to have 'standalone = false',
it does make sense that errors / headers / redirects from components that are not
part of the flushed output should still have their headers / errors etc added:
Otherwise, standalone should had been true in the first place. Also note that I did that is with async
component rendering with standalone = false : It works as expected.

Last thing I think that could be improved is that instead of directly invoking the constructor
 of HstServletResponseState and then bind the response state to the 'window' by invoking
 window.setResponseState(this), it might be nicer to have a method on the HstComponentWindow.
 Something like
 HstComponentWindow#bindResponseState(.....) where this method is responsible for creating the
 response state *and* set binds the window to the response state. That would imo be a cleaner approach
 which we could add later