HSTTWO-3623 Make sure the *caching* objectconverter is used in HippoTranslationBeanServiceImpl
Apart from making sure the *caching* objectconverter is used in HippoTranslationBeanServiceImpl,
I did cosmetic changes from making variables final where possible, update copyright years, remove
the imo redundant HippoTranslatedContentUtils static method and move the content into
HippoTranslationBeanServiceImpl#getTranslationNodes. Also in the AvailableTranslations bean
I deprecated the constructor with the ObjectConverter argument since this is not needed any more

Note that to be able to get hold of the object converter as used in ther HstRequestContext (which is in
general a cacheable objectconverter) I have added #getObjectConverter to the HstRequestContext interface.