HSTTWO-3743 Add an html comment as epilogue in case for  a component rendering request there are unprocessed head contributions
Via the HeadContributionsTag, we now keep track of which head contributions are really processed
in the top level response. When rendering the top level hstResponse(State), we can check whether there are
unprocessed head contributions. If the request is a component rendering request, we then add an html comments
to the response as follows:

<!-- unprocessed-head-elements-present  -->

Note that I removed some original code wrt 'addResponseHeadElement' : This method in effect *never* did add
anything because only when the 'parentResponse' is not an instance of HstResponse, the 'headElements' can be
 non-null