HSTTWO-3787 Only short-circuit on isActionResponse
Theoretically, for resource responses, we could in the future start supporting a (sub)tree
of components to be processed, hence, do not short circuit children flushing for resource responses.

At the current situation, a resource response will have possible child windows, but the response state of those
child windows will be null (at this moment). Since we now have the null check on the response state, this will be
no problem