HSTTWO-3792 Sanitize the FormField object to be better suited for json (de)serialization
It doesn't make sense that 'values' is a map where always the key and value are the same. It
should just be a list. For backwards compatibility, I keep the public methods but store the value in
'valueList' from now on. The #get/setValues should not be serialized (jackson) hence the @JsonIgnore

Also, it is easier for messages and valueList to be never null. No null checks needed and (de)serialization via
jackson object mapper is easier and more robust