HSTTWO-3882 When restoring primary cache entry from second level cache, use new creation time
When using the creation time of the second level cache entry, it can happen that the primary cache entry
is directly expired: Assume the second level cache entry got a creation time 2 minutes ago, and the newTTL
of the primary cache entry is 60 seconds (aka the second level cache entries live for 3 minutes). Then the
primary cache entry is directly expired. Instead, the primary cache entry should be stored with the
original key instance *and* with a TTL of 60 seconds and a creation time of NOW