HSTTWO-3991 Make sure to always use the corrent 'rendering mount'
When in the channel manager, requests like /_rp/ that hit a 'rest' endpoint do not
resolve to the 'editing mount' but to the 'rp mount'. When trying to find which
channel to render, we first need to check whether there is the static CMS_REQUEST_RENDERING_MOUNT_ID
on the http session. If there is, use that one to find out which mount to render