HSTTWO-3976 Support to create a channel from a blueprint that only has a hst:site node
In this case the hst:site node must point to some existing hst configuration node. The new
channel will result in an extra hst:configuration that inherits from the configuration where the
hst:site node points to. Also it will have a hst:channel node in its workspace to make sure the channel
is visible in the channel mngr