HSTTWO-4094 Make sure for a single Mount never two or more channels are returned
Instead of looping through the Mount objects, we get all the channels
directly from the VirtualHosts object. Via this methods, never duplicate
channels can be returned. When there is a preview and a live version of
a channel, we return the preview channel since this seems to be required
to make the channel mngr working correctly. Added integration tests
to confirm that this behavior (and cast the spec in concrete) and also
added an integration test that assures VirtualHosts#getChannels only returns
channels for the current webapp (context path).